+++ DarkKitchen.BusinessLogic/ShippingCosts/ShippingCost24hsCalculator.cs
@@ -4,8 +4,8 @@ namespace DarkKitchen.BusinessLogic.ShippingCosts;
 
 public class ShippingCost24hsCalculator : IShippingCostCalculator
 {
-    private readonly double _shippingCost24Hs = 10;
-    public double GetCost()
+    private readonly decimal _shippingCost24Hs = 10m;
+    public decimal GetCost()
     {
         return _shippingCost24Hs;
     }
